requirements.txt
flask
flask-cors
requests
python-dotenv
google-generativeai
spacy
deepface
opencv-python
tf-keras
tensorflow
